Fix false edit warning bug when creating own swp
Typing fast would trigger lots of document changed events which could give a false warning that the document was being edited elsewhere.

This happended if an event was processed while a previous one was still creating the file and the contents had not yet been written. If fs.stat could return EEXISTS but the contents could not be used to verify that the .swp file belonged to the client itself and the extension had to make the assumption that it was someone elses .swp file.